+++ backend/database/migrations/2021_06_01_195428_create_submissions_tables.php
@@ -62,7 +62,7 @@ class CreateSubmissionsTables extends Migration
         Schema::table('submissions', function (Blueprint $table) {
             $table->dropForeign('submissions_publication_id_foreign');
         });
-        Schema::dropIfExists('submissions_user');
+        Schema::dropIfExists('submission_user');
         Schema::dropIfExists('submissions');
     }
 }