+++ backend/database/seeders/PublicationSeeder.php
@@ -6,6 +6,7 @@ namespace Database\Seeders;
 use App\Models\Publication;
 use App\Models\StyleCriteria;
 use App\Models\User;
+use Faker\Factory as Faker;
 use Illuminate\Database\Seeder;
 
 class PublicationSeeder extends Seeder
@@ -17,6 +18,7 @@ class PublicationSeeder extends Seeder
      */
     public function run()
     {
+        $faker = Faker::create();
         $this->callOnce(UserSeeder::class);
 
         Publication::factory()
@@ -25,6 +27,8 @@ class PublicationSeeder extends Seeder
         ->create([
             'id' => 1,
             'name' => 'CCR Test Publication 1',
+            'home_page_content' => $faker->paragraphs(2, true),
+            'new_submission_content' => $faker->paragraphs(2, true),
         ]);
 
         Publication::factory()