+++ client/src/components/atoms/SubmissionContent.vue
@@ -52,29 +52,34 @@
     <h1>Grouped Avatars</h1>
     <sample-submission-content-avatars />
   </article>
+  <article
+    ref="contentRef"
+    data-cy="content"
+    class="col-sm-9 submission-content"
+  >
+    <bubble-menu
+      v-if="editor"
+      :editor="editor"
+      :tippy-options="{ duration: 100 }"
+      :should-show="shouldShow"
+    >
+      <q-btn color="white" text-color="primary" @click="addComment">
+        <q-icon name="add_comment" />
+      </q-btn>
+    </bubble-menu>
+    <editor-content :editor="editor" />
+  </article>
 </template>
+
 <script setup>
-import { ref } from "vue"
-import { Editor } from "@tiptap/vue-3"
+import { ref, inject, computed } from "vue"
+import { Editor, EditorContent, BubbleMenu } from "@tiptap/vue-3"
 import Highlight from "@tiptap/extension-highlight"
 import StarterKit from "@tiptap/starter-kit"
 import SampleSubmissionContent from "./SampleSubmissionContent.vue"
 import SampleSubmissionContentGrouped from "./SampleSubmissionContentGrouped.vue"
 import SampleSubmissionContentCriteria from "./SampleSubmissionContentCriteria.vue"
 import SampleSubmissionContentAvatars from "./SampleSubmissionContentAvatars.vue"
-  <article
-    ref="contentRef"
-    data-cy="content"
-    class="col-sm-9 submission-content"
-  >
-    <editor-content :editor="editor" />
-  </article>
-</template>
-<script setup>
-import { ref, inject, computed } from "vue"
-import { Editor, EditorContent } from "@tiptap/vue-3"
-import Highlight from "@tiptap/extension-highlight"
-import StarterKit from "@tiptap/starter-kit"
 import AnnotationExtension from "src/tiptap/annotation-extension"
 
 const submission = inject("submission")
@@ -139,7 +144,27 @@ const editor = new Editor({
     AnnotationExtension.configure({ annotations }),
   ],
 })
-console.log(typeof editor)
+
+function shouldShow({ state }) {
+  return !state.selection.empty
+}
+
+function addComment() {
+  const [from, to] = [
+    editor.state.selection.$anchor.pos,
+    editor.state.selection.$head.pos,
+  ].sort((a, b) => a - b)
+  const range = { from, to }
+  activeComment.value = {
+    __typename: "InlineComment",
+    new: true,
+    from,
+    to,
+    parent_id: null,
+    id: "new",
+  }
+  console.log(range)
+}
 </script>
 
 <style lang="scss">